[FIX] product_pricelist_supplierinfo: inherit the _compute_price method in product_pricelist model
In version 15.0 the _compute_price_rule method was used to obtain the
product price because the price was a computed field, however, in
version 16.0 we need to inherit _compute_price from product_pricelist_item
because this method is used directly when creating a new sale order line.
