+++ product_pricelist_supplierinfo/models/__init__.py
@@ -1,6 +1,6 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from . import product_supplierinfo
-from . import product_pricelist
+from . import product_pricelist_item
 from . import product_product
 from . import product_template
